Hotfix VS incorrectly changing focus after navgation (#75)
This PR hotfixes an issue where Visual Studio incorrectly focuses the
Editor Bar after changing the active tab.

The fix ensures that when focus lands on the first tab navigation
control in the Editor Bar without coming from its neighbor, the focus
change is ignored, and focus is redirected to the text view instead.

Fixed: #74